Modifications to validated and age fields in evidence page
