+++ client/src/components/Evidences.js
@@ -1,38 +1,7 @@
 import React, {Component} from 'react';
 import {Icon, Table, Popover} from 'antd';
 import BioID from './tags/BioID';
-
-const getIcon = (mess) => {
-	if (mess === "Yes" || mess === "YES") return "check-circle"
-	if (mess === "No" || mess === "NO") return "stop"
-	if (mess === "Potential" || mess === "POTENTIAL") return "info-circle"
-	return "minus-circle"
-}
-
-const getColor = (mess) => {
-	if (mess === "Yes" || mess === "YES") return "#52c41a"
-	if (mess === "No" || mess === "NO") return "red"
-	if (mess === "Potential" || mess === "POTENTIAL") return "blue"
-	return "gray"
-}
-
-const getPopover = (mess) => {
-	if (mess === "Yes" || mess === "YES") return "Yes"
-	if (mess === "No" || mess === "NO") return "No"
-	if (mess === "Potential" || mess === "POTENTIAL") return "Potential"
-	return "Unavailable"
-}
-
-const renderIcon = (val) => {
-	return <Popover content={getPopover(val)}>
-		<Icon 
-		type={getIcon(val)}
-		theme="twoTone"
-		twoToneColor={getColor(val)}
-		style={{ fontSize: '16px' }} />
-  	</Popover>
-}
-
+import Marker from './../Utilities/Marker';
 const columns = [{
 	title: 'ID',
 	dataIndex: 'id',
@@ -44,27 +13,37 @@ const columns = [{
 	title: 'Validated',
 	dataIndex: 'application.validated',
 	key: 'validated',
-	render: val => renderIcon(val)
+	render: val => (
+		<Marker val={val}></Marker> 
+	)	
 },{
 	title: 'Diagnosis',
 	dataIndex: 'application.diagnosis',
 	key: 'diagnosis',
-	render: val => renderIcon(val)
+	render: val => (
+		<Marker val={val}></Marker> 
+	)	
 },{
 	title: 'Predictive',
 	dataIndex: 'application.predictive',
 	key: 'predictive',
-	render: val => renderIcon(val)
+	render: val => (
+		<Marker val={val}></Marker> 
+	)	
 },{
 	title: 'Prognostis',
 	dataIndex: 'application.prognosis',
 	key: 'prognosis',
-	render: val => renderIcon(val)
+	render: val => (
+		<Marker val={val}></Marker> 
+	)	
 },{
 	title: 'Therapeutic',
 	dataIndex: 'application.therapeutic',
 	key: 'therapeutic',
-	render: val => renderIcon(val)
+	render: val => (
+		<Marker val={val}></Marker> 
+	)	
 }];
 
 class Evidences extends Component {
